fix basic functionality being broken

+++ bot/bot.py
@@ -447,56 +447,51 @@ async def on_message(message: discord.Message):
                     eventType = type(e).__name__, trace = traceback.format_exc())
 
     # Non-command messages
-    elif not isDM:
-        callingGuild = botState.guildsDB.getGuild(message.guild.id)
-        if message.channel.id == callingGuild.storyChannelID:
-            if message.author.id == callingGuild.lastAuthorID:
-                await message.channel.send(":boom: **Story broken, " + message.author.mention + "!** It wasn't your turn!")
+    elif not isDM and (callingGuild := botState.guildsDB.getGuild(message.guild.id)).storyChannelID == message.channel.id:
+        if message.author.id == callingGuild.lastAuthorID:
+            await message.channel.send(":boom: **Story broken, " + message.author.mention + "!** It wasn't your turn!")
+            if callingGuild.story:
+                await message.channel.send(callingGuild.story)
+            callingGuild.story = ""
+            callingGuild.lastAuthorID = -1
+            return
+        elif message.content == ".":
+            if callingGuild.story != "":
+                await message.channel.send("**Story complete!**")
+                await message.channel.send(callingGuild.story)
+                callingGuild.story = ""
+                callingGuild.lastAuthorID = -1
+        else:
+            msgSplit = message.content.split(" ")
+            if len(msgSplit) > 2 or len(msgSplit) == 2 and msgSplit[0] not in cfg.ignoredSymbols:
+                await message.channel.send(f":boom: **Story broken, {message.author.mention}!** You only add one {'emoji' if callingGuild.emojiOnly else 'word'} at a time!")
                 if callingGuild.story:
                     await message.channel.send(callingGuild.story)
                 callingGuild.story = ""
                 callingGuild.lastAuthorID = -1
-                return
-
-            if callingGuild.emojiOnly:
-                if message.content == ".":
-                    await message.channel.send("**Story complete!**")
-            elif message.content == "" or message.content == "." and callingGuild.story == "":
-                pass
-
-            elif " " in message.content:
-                firstWord = message.content.split(" ")[0]
-                if len(message.content.split(" ")) > 2 or not(firstWord == "..." or len(firstWord) == 1 and firstWord in cfg.ignoredSymbols):
-                    await message.channel.send(":boom: **Story broken, " + message.author.mention + "!** You only add one " + ("emoji" if callingGuild.emojiOnly else "word") + " at a time!")
-                    if callingGuild.story:
-                        await message.channel.send(callingGuild.story)
-                    callingGuild.story = ""
-                    callingGuild.lastAuthorID = -1
-                    if callingGuild.emojiOnlyErrSent:
-                        callingGuild.emojiOnlyErrSent = False
-                elif not (lib.emojis.strIsCustomEmoji(message.content) or lib.emojis.strIsUnicodeEmoji(message.content)):
+                if callingGuild.emojiOnlyErrSent:
+                    callingGuild.emojiOnlyErrSent = False
+            elif len(callingGuild.story) + len(message.content) > 2000:
+                await message.channel.send(":boom: **Max story length exceeded!**")
+                await message.channel.send(callingGuild.story)
+                callingGuild.story = ""
+                callingGuild.lastAuthorID = -1
+            elif callingGuild.emojiOnly and not (lib.emojis.strIsCustomEmoji(message.content) or lib.emojis.strIsUnicodeEmoji(message.content)):
+                await message.delete()
+                if not callingGuild.emojiOnlyErrSent:
+                    await message.channel.send(message.author.mention + " emoji only mode is enabled! You can only contribute a single emoji.")
+                    callingGuild.emojiOnlyErrSent = True
+            else:
+                try:
+                    lib.emojis.BasedEmoji.fromStr(message.content, rejectInvalid=True)
+                except (lib.exceptions.UnrecognisedCustomEmoji, TypeError):
                     await message.delete()
-                    if not callingGuild.emojiOnlyErrSent:
-                        await message.channel.send(message.author.mention + " emoji only mode is enabled! You can only contribute a single emoji.")
-                        callingGuild.emojiOnlyErrSent = True
+                    await message.channel.send(message.author.mention + " please only use standard emojis or emojis from this server!")
                 else:
-                    try:
-                        lib.emojis.BasedEmoji.fromStr(message.content, rejectInvalid=True)
-                    except lib.exceptions.UnrecognisedCustomEmoji:
-                        await message.channel.send(message.author.mention + " please only use standard emojis or emojis from this server!")
-                        return
                     if callingGuild.emojiOnlyErrSent:
                         callingGuild.emojiOnlyErrSent = False
-
-                    if len(callingGuild.story) + len(message.content) > 2000:
-                        await message.channel.send(":boom: **Max story length exceeded!**")
-                        await message.channel.send(callingGuild.story)
-                        callingGuild.story = ""
-                        callingGuild.lastAuthorID = -1
-                    else:
-                        callingGuild.story += message.content
-                        callingGuild.lastAuthorID = message.author.id
-                
+                    callingGuild.story += message.content
+                    callingGuild.lastAuthorID = message.author.id
 
 
 @botState.client.event
